requirements.txt
beautifulsoup4==4.3.2
functools32==3.2.3-1
python-Levenshtein==0.12.0
python-dateutil==2.3
requests==2.5.1
six==1.8.0
wsgiref==0.1.2
